Move internal signal classes into internal sub package
