+++ code/android-signal-library/src/main/java/com/adobe/marketing/mobile/signal/internal/SignalHitProcessor.kt
@@ -9,7 +9,7 @@
   governing permissions and limitations under the License.
 */
 
-package com.adobe.marketing.mobile.signal
+package com.adobe.marketing.mobile.signal.internal
 
 import androidx.annotation.VisibleForTesting
 import com.adobe.marketing.mobile.services.DataEntity
